Moving main threads to /threads
