+++ backend/__init__.py
@@ -2,13 +2,13 @@
 # Import modules
 
 from .server              import ServerThread
-from .app                 import APPThread
+from .threads.app         import APPThread
 
-from .can                 import CANThread
-from .swc                 import SWCThread
-from .adc                 import ADCThread
-from .rti                 import RTIThread
-from .ign                 import IGNThread
+from .threads.can         import CANThread
+from .threads.swc         import SWCThread
+from .threads.adc         import ADCThread
+from .threads.rti         import RTIThread
+from .threads.ign         import IGNThread
 
 
 from .dev.vcan            import VCANThread
